Update README.md
Welcome to the Hacktober Fest Website Project! In this project, you will have the opportunity to apply your newly acquired HTML and CSS skills to create a website based on the theme of Hacktober Fest. Hacktober Fest is a month-long celebration of open-source software, and your task is to design a website that provides information about the event, its goals, and how participants can get involved.